Moved DuplicateCourseRedirect concern
