+++ app/controllers/candidate_interface/continuous_applications/course_choices/course_site_controller.rb
@@ -2,7 +2,7 @@ module CandidateInterface
   module ContinuousApplications
     module CourseChoices
       class CourseSiteController < BaseController
-        include Concerns::DuplicateCourseRedirect
+        include CandidateInterface::CourseChoices::Concerns::DuplicateCourseRedirect
 
       private
